qna detail 수정 -박수연

+++ src/app/notice/qna-detail/qna-detail.component.ts
@@ -25,18 +25,18 @@ export class QnaDetailComponent implements OnInit {
   ngOnInit() {
     this.qna = this.qnaService.getQnaObject();
 
-    if(this.qna == null){
+    if (this.qna == null) {
       var q_no = this.tokenService.getToken("qnaDetailToken")
       var qnaList = this.tokenService.getToken("qnaToken");
-      var qna = qnaList.find(function (item) {
+      var qna = qnaList.find(function(item) {
         return item.q_no === q_no;
       });
       var q_content = qna.q_content;
-      qna.q_content = qna.q_content.replace("\r\n","<br>");
+      qna.q_content = qna.q_content.replace("\r\n", "<br>");
       this.qna = qna;
     }
 
-    if(this.qna.q_reply == this.reply) {
+    if (this.qna.q_reply == this.reply) {
       this.reply = 'N';
     } else {
       this.reply = "답변 완료"
@@ -48,11 +48,16 @@ export class QnaDetailComponent implements OnInit {
   }
 
   deleteQna(q_no: number) {
-    return this.qnaService.deleteQna(q_no)
-      .subscribe(() => {
-        alert("고객님의 질문이 삭제되었습니다.");
-        // this.tokenService.removeToken("qnaToken");
-        this.router.navigate(["notice/qna"]);
-      });
+    var confirm = window.confirm('질문을 삭제하시겠습니까?');
+    if (confirm) {
+      return this.qnaService.deleteQna(q_no)
+        .subscribe(() => {
+          alert("고객님의 질문이 삭제되었습니다.");
+          // this.tokenService.removeToken("qnaToken");
+          this.router.navigate(["notice/qna"]);
+        });
+    } else {
+      return;
+    }
   }
 }
